LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-scroll-snap/input/keyboard-snap-interruption-expected.txt


FAIL Interrupting a snap animation with another key press targeting the same node preserves snapping destination assert_equals: Should successfully snap to the target despite interruption expected 600 but got 56
